Object serialization

+++ src/main/java/com/reggiemcdonald/neural/res/NumberImage.java
@@ -2,11 +2,11 @@ package com.reggiemcdonald.neural.res;
 
 public class NumberImage {
     public double[][] image;
-    public int[]      label;
+    public double[] label;
 
     public NumberImage (double[][] image, int label) {
         this.image = image;
-        this.label = new int[10];
+        this.label = new double[10];
         this.label [label] = 1;
     }
 }
